Implementing correct run functions.
In accelstepper lib, if you call a run() the user only needs to define an accelleration and a destination (move or moveTo).  For a runSpeed, set a maxSpeed and speed. For a runSpeedToPosition, define maxSpeed, speed and destination (move/moveTo). Currently these example are not working exactly as planned. Also broke the limit switch examples....